fix(output): set aws region as start mode output
This is useful when caller workflow provides a profile for stop
mode and if associated PR tends to be merged before instance
teardown. In that case Slab will try to fetch the remote
configuration file (slab.toml) on a reference that has vanished
thus resulting in an error in Slab without terminating the
instance.
Using the region as input for stop mode avoid trying to fetch
slab.toml file.