feat(slab): wait slab until async task is over
This is done to handle various timeouts configuration on spawn
request in Slab. Until now the timeout was set to 15 minutes. This
was fine since default spawn timeout in Slab was set to 10 minutes.
However in some cases, slab backend profiles can request a longer
duration before timeout.
Since the asynchronous task created in Slab will eventually
finish, we now simply wait for this moment.